+++ Backend/src/main/java/com/inn/cafe/serviceImpl/ProductServiceImpl.java
@@ -198,7 +198,7 @@ public class ProductServiceImpl implements ProductService {
         product.setCategory(category);
         product.setName(requestMap.get("name"));
         product.setDescription(requestMap.get("description"));
-        product.setPrice(Integer.parseInt(requestMap.get("price")));
+        product.setPrice(Double.parseDouble(requestMap.get("price")));
         return product;
     }
